pyproject.toml
[project]
name = "qwen-experiments"
version = "0.1.0"
description = "Stream-first runner bakeoff for Qwen TTS outside the main Aether Voice repo."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "orjson>=3.10.18",
  "pydantic>=2.11.3",
  "python-dotenv>=1.1.0",
  "soundfile>=0.13.1",
  "tenacity>=9.1.2",
  "typer>=0.16.1",
  "uvicorn>=0.34.0",
  "websockets>=15.0.1",
]

[project.optional-dependencies]
native = [
  "qwen-tts",
]
vllm = [
  "vllm-omni",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qwen_experiments"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
default-groups = ["dev"]
